Added buttons to change the quantity of items in pantry and removed some unnecessary commented code from demo forms.

+++ JP_PantryAid/PantryAid.Core/Interfaces/iIngredientData.cs
@@ -18,11 +18,11 @@ namespace PantryAid.Core.Interfaces
         Ingredient GetIngredient(int ID);
         Ingredient GetIngredient(string name);
         List<IngredientItem> GetIngredientsFromPantry(int PantryID);
-        int AddIngredientToPantry(int PantryID, int IngredientID, float Quantity = 1.0f);
-        int AddIngredientToPantry(int PantryID, Ingredient ingredient, float Quantity = 1.0f);
+        int AddIngredientToPantry(int PantryID, int IngredientID, double Quantity = 1.0f);
+        int AddIngredientToPantry(int PantryID, Ingredient ingredient, double Quantity = 1.0f);
         int RemoveIngredientFromPantry(int PantryID, int IngredientID);
         int RemoveIngredientFromPantry(int PantryID, Ingredient ingredient);
-        int UpdatePantryIngredientQuantity(int PantryID, int IngredientID, float NewQuantity);
+        int UpdatePantryIngredientQuantity(int PantryID, int IngredientID, double NewQuantity);
         int AddIngredientsFromRecipeShort(Recipe_Short recipe);
         int AddIngredientsFromRecipeFull(Recipe_Full recipe);
     }
